requirements.txt
Django>=4.2,<5.0
django-cms>=4.1
djangocms-frontend>=1.3
djangocms-text-ckeditor>=5.1
django-filer>=3.0
easy-thumbnails>=2.8
djangocms-alias>=2.0
djangocms-versioning>=2.0
Pillow>=10.0
psycopg2-binary>=2.9
whitenoise>=6.6
django-sekizai>=4.0
